fix : 게시글 등록 수정
- 관리자로그인 게시글 등록 코드 수정 (springboot & react)

+++ playground/src/main/java/com/players/playground/Board_Notice/repository/NoticeImageRepository.java
@@ -3,7 +3,7 @@ package com.players.playground.Board_Notice.repository;
 import com.players.playground.Board_Notice.entity.NoticeImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeImageRepository extends JpaRepository <NoticeImage, Long> {
+public interface NoticeImageRepository extends JpaRepository <NoticeImage, Integer> {
 
 
 }
